plot 10 models

+++ tools/visualisation.py
@@ -1,11 +1,12 @@
 import os
 import pandas as pd
-
 import matplotlib
+#matplotlib.use('Agg')
 matplotlib.use('TkAgg')
 import matplotlib.pyplot as plt
 import numpy as np
 from timeit import default_timer as timer
+import matplotlib.gridspec as gridspec
 
 
 def plot_real_wind(cf):
@@ -51,22 +52,23 @@ def plt_forecast_wind_train(cf):
             wind_real_df_model_day = wind_real_df_model.loc[wind_real_df_model['date_id'] == d_unique]
             hour_unique = sorted(pd.unique(wind_real_df_model_day['hour']))
             for h_unique in hour_unique:
-                print('Processing forecast data for model: %d,  date: %d, hour: %d' % (m_unique, d_unique, h_unique))
-                wind_real_df_model_day_hour = wind_real_df_model_day.loc[wind_real_df_model_day['hour'] == h_unique]
+                if not os.path.exists(os.path.join(cf.wind_save_path, 'Train_forecast_wind_model_%d_day_%d_hour_%d.npy' % (m_unique, d_unique, h_unique))):
+                    print('Processing forecast data for model: %d,  date: %d, hour: %d' % (m_unique, d_unique, h_unique))
+                    wind_real_df_model_day_hour = wind_real_df_model_day.loc[wind_real_df_model_day['hour'] == h_unique]
 
-                if not len(x_unique) * len(y_unique) == wind_real_df_model_day_hour.index.__len__():
-                    print('There are some missing data or redudant data: pixel range: %d, given wind pixel range: %d.' %
-                          (len(x_unique) * len(y_unique), wind_real_df_model_day_hour.index.__len__()))
+                    if not len(x_unique) * len(y_unique) == wind_real_df_model_day_hour.index.__len__():
+                        print('There are some missing data or redudant data: pixel range: %d, given wind pixel range: %d.' %
+                              (len(x_unique) * len(y_unique), wind_real_df_model_day_hour.index.__len__()))
 
-                wind_real_day_hour = np.zeros(shape=(len(x_unique), len(y_unique)))
-                for idx in range(wind_real_df_model_day_hour.index.__len__()):
-                    x_loc = int(wind_real_df_model_day_hour.iloc[idx]['xid']) - 1
-                    y_loc = int(wind_real_df_model_day_hour.iloc[idx]['yid']) - 1
-                    wind = int(wind_real_df_model_day_hour.iloc[idx]['wind'])
-                    wind_real_day_hour[x_loc, y_loc] = wind
+                    wind_real_day_hour = np.zeros(shape=(len(x_unique), len(y_unique)))
+                    for idx in range(wind_real_df_model_day_hour.index.__len__()):
+                        x_loc = int(wind_real_df_model_day_hour.iloc[idx]['xid']) - 1
+                        y_loc = int(wind_real_df_model_day_hour.iloc[idx]['yid']) - 1
+                        wind = int(wind_real_df_model_day_hour.iloc[idx]['wind'])
+                        wind_real_day_hour[x_loc, y_loc] = wind
 
-                np.save(os.path.join(cf.wind_save_path, 'Train_forecast_wind_model_%d_day_%d_hour_%d.npy' % (m_unique, d_unique, h_unique)),
-                        wind_real_day_hour)
+                    np.save(os.path.join(cf.wind_save_path, 'Train_forecast_wind_model_%d_day_%d_hour_%d.npy' % (m_unique, d_unique, h_unique)),
+                            wind_real_day_hour)
 
 
 def plt_forecast_wind_test(cf):
@@ -108,33 +110,161 @@ def plot_all_wind(cf):
     # Create the data generators
     city_data_df = pd.read_csv(os.path.join(cf.dataroot_dir, 'CityData.csv'))
 
-    weather_list = sorted(os.listdir(cf.savepath), reverse=True)
-    for weather_name in weather_list:
-        wind_real_day_hour = np.load(os.path.join(cf.savepath, weather_name))
+    if cf.plot_train_model:
         # plot figures here
-        #plt.close()
+        fig = plt.figure(num=1)
+        fig.clf()
+        mng = plt.get_current_fig_manager()
+        mng.resize(*mng.window.maxsize())
+        left = gridspec.GridSpec(1, 1)
+        left.update(left=0.05, right=0.2, hspace=0.05)
+        right = gridspec.GridSpec(2, 5)
+        right.update(left=0.25, right=0.98, hspace=0.05)
+
+        total_weather = np.zeros(shape=(cf.grid_world_shape[0], cf.grid_world_shape[1], len(cf.model_unique)+1))
+        for d_unique in cf.day_list:
+            for h_unique in range(cf.hour_unique[0], cf.hour_unique[1]+1):
+                print('Processing forecast data for  date: %d, hour: %d' % (d_unique, h_unique))
+                wind_real_day_hour = np.load(os.path.join(cf.wind_save_path, 'real_wind_day_%d_hour_%d.npy'%(d_unique, h_unique)))
+                total_weather[:, :, 0] = wind_real_day_hour
+                for m_unique in cf.model_unique:
+                    np_file = os.path.join(cf.wind_save_path, 'Train_forecast_wind_model_%d_day_%d_hour_%d.npy' % (m_unique, d_unique, h_unique))
+                    total_weather[:, :, m_unique] = np.load(np_file)
+
+                # We need to normalise the map to standardize visualisation
+                total_weather_max = np.max(total_weather)
+                total_weather = (total_weather-15)/total_weather_max
+                # plot all the real and forecast model here:
+                fig.clf()
+                plt.subplot(left[:, :])
+                plt.imshow(total_weather[:, :, 0],  cmap='jet')
+                plt.colorbar()
+                plt.title('real_wind_day_%d_hour_%d' % (d_unique, h_unique))
+                # we also plot the city location
+                for idx in range(city_data_df.index.__len__()):
+                    x_loc = int(city_data_df.iloc[idx]['xid']) - 1
+                    y_loc = int(city_data_df.iloc[idx]['yid']) - 1
+                    cid = int(city_data_df.iloc[idx]['cid'])
+                    plt.scatter(y_loc, x_loc, c='r', s=40)
+                    plt.annotate(str(cid), xy=(y_loc, x_loc), color='white', fontsize=20)
+
+                # we also draw some contours
+                x = np.arange(0, wind_real_day_hour.shape[1], 1)
+                y = np.arange(0, wind_real_day_hour.shape[0], 1)
+                X, Y = np.meshgrid(x, y)
+                CS = plt.contour(X, Y, total_weather[:, :, 0], (0,), colors='k')
+                plt.clabel(CS, inline=1, fontsize=10)
+                for m_unique in cf.model_unique:
+                    # plot all the real and forecast model here:
+                    plt.subplot(right[(m_unique-1)//5, np.mod(m_unique, 5)-1])
+                    plt.imshow(total_weather[:, :, m_unique],  cmap='jet')
+                    # we also plot the city location
+                    for idx in range(city_data_df.index.__len__()):
+                        x_loc = int(city_data_df.iloc[idx]['xid']) - 1
+                        y_loc = int(city_data_df.iloc[idx]['yid']) - 1
+                        cid = int(city_data_df.iloc[idx]['cid'])
+                        plt.scatter(y_loc, x_loc, c='r', s=40)
+                        plt.annotate(str(cid), xy=(y_loc, x_loc), color='white', fontsize=20)
+
+                    # we also draw some contours
+                    x = np.arange(0, wind_real_day_hour.shape[1], 1)
+                    y = np.arange(0, wind_real_day_hour.shape[0], 1)
+                    X, Y = np.meshgrid(x, y)
+                    CS = plt.contour(X, Y, total_weather[:, :, m_unique], (0,), colors='k')
+                    plt.clabel(CS, inline=1, fontsize=10)
+                    plt.title('Model %d' % m_unique)
+
+                mng = plt.get_current_fig_manager()
+                mng.resize(*mng.window.maxsize())
+                plt.waitforbuttonpress(0.0001)
+                # plt.show()
+                save_fig_name = os.path.join(cf.fig_save_train_path, '%s.png' % ('Train_real_models_wind_day_%d_hour_%d' % (d_unique, h_unique)))
+                plt.savefig(save_fig_name, dpi=74, bbox_inches='tight')
+
+    elif cf.plot_test_model:
+        # plot figures here
+        fig = plt.figure(num=1)
+        fig.clf()
+        mng = plt.get_current_fig_manager()
+        mng.resize(*mng.window.maxsize())
+        right = gridspec.GridSpec(2, 5)
+        right.update(left=0.02, right=0.98, hspace=0.05)
+
+        total_weather = np.zeros(shape=(cf.grid_world_shape[0], cf.grid_world_shape[1], len(cf.model_unique)))
+        for d_unique in cf.day_list:
+            for h_unique in range(cf.hour_unique[0], cf.hour_unique[1] + 1):
+                print('Processing forecast data for  date: %d, hour: %d' % (d_unique, h_unique))
+                for m_unique in cf.model_unique:
+                    np_file = os.path.join(cf.wind_save_path, 'Test_forecast_wind_model_%d_day_%d_hour_%d.npy' % (
+                    m_unique, d_unique, h_unique))
+                    total_weather[:, :, m_unique-1] = np.load(np_file)
+
+                # We need to normalise the map to standardize visualisation
+                fig.clf()
+                total_weather_max = np.max(total_weather)
+                total_weather = (total_weather - 15) / total_weather_max
+
+                for m_unique in cf.model_unique:
+                    # plot all the real and forecast model here:
+                    plt.subplot(right[(m_unique - 1) // 5, np.mod(m_unique, 5) - 1])
+                    plt.imshow(total_weather[:, :, m_unique-1], cmap='jet')
+                    # we also plot the city location
+                    for idx in range(city_data_df.index.__len__()):
+                        x_loc = int(city_data_df.iloc[idx]['xid']) - 1
+                        y_loc = int(city_data_df.iloc[idx]['yid']) - 1
+                        cid = int(city_data_df.iloc[idx]['cid'])
+                        plt.scatter(y_loc, x_loc, c='r', s=40)
+                        plt.annotate(str(cid), xy=(y_loc, x_loc), color='white', fontsize=20)
+
+                    # we also draw some contours
+                    x = np.arange(0, cf.grid_world_shape[1], 1)
+                    y = np.arange(0, cf.grid_world_shape[0], 1)
+                    X, Y = np.meshgrid(x, y)
+                    CS = plt.contour(X, Y, total_weather[:, :, m_unique-1], (0,), colors='k')
+                    plt.clabel(CS, inline=1, fontsize=10)
+
+                mng = plt.get_current_fig_manager()
+                mng.resize(*mng.window.maxsize())
+                fig.suptitle('Test_forecast_wind_day_%d_hour_%d.npy' % (d_unique, h_unique), size=20)
+                plt.waitforbuttonpress(0.0001)
+                save_fig_name = os.path.join(cf.fig_save_test_path, '%s.png' % ('Test_models_wind_day_%d_hour_%d' % (d_unique, h_unique)))
+                plt.savefig(save_fig_name, dpi=74, bbox_inches='tight')
+
+    else:
         plt.figure(1)
         plt.clf()
-        plt.imshow(wind_real_day_hour, cmap='jet')
-        plt.colorbar()
-        # we also plot the city location
-        for idx in range(city_data_df.index.__len__()):
-            x_loc = int(city_data_df.iloc[idx]['xid']) - 1
-            y_loc = int(city_data_df.iloc[idx]['yid']) - 1
-            cid = int(city_data_df.iloc[idx]['cid'])
-            plt.scatter(y_loc, x_loc, c='r', s=40)
-            plt.annotate(str(cid), xy=(y_loc, x_loc), color='white', fontsize=20)
-
-        # we also draw some contours
-        x = np.arange(0, wind_real_day_hour.shape[1], 1)
-        y = np.arange(0, wind_real_day_hour.shape[0], 1)
-        X, Y = np.meshgrid(x, y)
-        CS = plt.contour(X, Y, wind_real_day_hour,  (15,), colors='k')
-
-        plt.clabel(CS, inline=1, fontsize=10)
-        plt.title(weather_name[:-7])
 
         mng = plt.get_current_fig_manager()
         mng.resize(*mng.window.maxsize())
-        plt.waitforbuttonpress(0.0001)
-        plt.savefig(os.path.join(cf.fig_save_path, '%s.png'%(weather_name[:-7])), dpi=74, bbox_inches='tight')
+        # we plot everything
+        weather_list = sorted(os.listdir(cf.wind_save_path), reverse=True)
+        for weather_name in weather_list:
+            wind_real_day_hour = np.load(os.path.join(cf.wind_save_path, weather_name))
+
+            plt.clf()
+            plt.imshow(wind_real_day_hour, cmap='jet')
+            plt.colorbar()
+            # we also plot the city location
+            for idx in range(city_data_df.index.__len__()):
+                x_loc = int(city_data_df.iloc[idx]['xid']) - 1
+                y_loc = int(city_data_df.iloc[idx]['yid']) - 1
+                cid = int(city_data_df.iloc[idx]['cid'])
+                plt.scatter(y_loc, x_loc, c='r', s=40)
+                plt.annotate(str(cid), xy=(y_loc, x_loc), color='white', fontsize=20)
+
+            # we also draw some contours
+            x = np.arange(0, wind_real_day_hour.shape[1], 1)
+            y = np.arange(0, wind_real_day_hour.shape[0], 1)
+            X, Y = np.meshgrid(x, y)
+            CS = plt.contour(X, Y, wind_real_day_hour,  (15,), colors='k')
+
+            plt.clabel(CS, inline=1, fontsize=10)
+            plt.title(weather_name[:-4])
+
+            mng = plt.get_current_fig_manager()
+            mng.resize(*mng.window.maxsize())
+            plt.waitforbuttonpress(0.0001)
+            #plt.show()
+            save_fig_name = os.path.join(cf.fig_save_path, '%s.png'% (weather_name[:-4]))
+            print('Saving figure %s' % save_fig_name)
+            plt.savefig(save_fig_name, dpi=74, bbox_inches='tight')
